Diagnostics Tool for HETU

Produce a data frame of PINs that may require closer scrutiny.

Usage

hetu_diagnostic(pin, extract = NULL, subsetting = FALSE)

pin_diagnostic(pin, extract = NULL, subsetting = FALSE)

Arguments

pin

Finnish personal identification number as a character vector, or vector of identification numbers as a character vectors

extract

Extract only selected part of the diagnostic information. Valid values are "hetu", "is.temp", "valid.p.num", "valid.ctrl.char", "correct.ctrl.char", "valid.date", "valid.day", "valid.month", "valid.length", "valid.century". If NULL (default), returns all information.

subsetting

Print only PINs where the validity check chosen in extract returns FALSE.

Value

A data.frame containing diagnostic checks about PINs.

Examples

diagnosis_example <- c("010101-0102", "111111-111Q", 
"010101B0101", "320101-0101", "011301-0101", 
"010101-01010", "010101-0011")
## Print all diagnoses
hetu_diagnostic(diagnosis_example)
#>           hetu is.temp valid.p.num valid.ctrl.char correct.ctrl.char valid.date
#> 1  010101-0102   FALSE        TRUE            TRUE             FALSE       TRUE
#> 2  111111-111Q   FALSE        TRUE           FALSE             FALSE       TRUE
#> 3  010101B0101   FALSE        TRUE            TRUE              TRUE      FALSE
#> 4  320101-0101   FALSE        TRUE            TRUE              TRUE      FALSE
#> 5  011301-0101   FALSE        TRUE            TRUE             FALSE      FALSE
#> 6 010101-01010   FALSE        TRUE            TRUE              TRUE       TRUE
#> 7  010101-0011   FALSE       FALSE            TRUE             FALSE       TRUE
#>   valid.day valid.month valid.year valid.length valid.century
#> 1      TRUE        TRUE       TRUE         TRUE          TRUE
#> 2      TRUE        TRUE       TRUE         TRUE          TRUE
#> 3      TRUE        TRUE       TRUE         TRUE         FALSE
#> 4     FALSE        TRUE       TRUE         TRUE          TRUE
#> 5      TRUE       FALSE       TRUE         TRUE          TRUE
#> 6      TRUE        TRUE       TRUE        FALSE          TRUE
#> 7      TRUE        TRUE       TRUE         TRUE          TRUE
# Extract century-related checks
hetu_diagnostic(diagnosis_example, extract = "valid.century")
#>           hetu valid.century
#> 1  010101-0102          TRUE
#> 2  111111-111Q          TRUE
#> 3  010101B0101         FALSE
#> 4  320101-0101          TRUE
#> 5  011301-0101          TRUE
#> 6 010101-01010          TRUE
#> 7  010101-0011          TRUE
# Extract only rows where valid.ctrl.char = FALSE
hetu_diagnostic(diagnosis_example, subsetting = TRUE, extract = "valid.day") 
#>          hetu    sex p.num ctrl.char date day month year century valid.pin
#> 1 320101-0101 Female   010         1 <NA>  NA     1 1901       -     FALSE
#>   valid.p.num valid.ctrl.char correct.ctrl.char valid.date valid.day
#> 1        TRUE            TRUE              TRUE      FALSE     FALSE
#>   valid.month valid.year valid.length valid.century is.temp
#> 1        TRUE       TRUE         TRUE          TRUE   FALSE

diagnosis_example <- c("010101-0102", "111111-111Q", 
"010101B0101", "320101-0101", "011301-0101", 
"010101-01010", "010101-0011")
## Print all diagnoses
pin_diagnostic(diagnosis_example)
#>           hetu is.temp valid.p.num valid.ctrl.char correct.ctrl.char valid.date
#> 1  010101-0102   FALSE        TRUE            TRUE             FALSE       TRUE
#> 2  111111-111Q   FALSE        TRUE           FALSE             FALSE       TRUE
#> 3  010101B0101   FALSE        TRUE            TRUE              TRUE      FALSE
#> 4  320101-0101   FALSE        TRUE            TRUE              TRUE      FALSE
#> 5  011301-0101   FALSE        TRUE            TRUE             FALSE      FALSE
#> 6 010101-01010   FALSE        TRUE            TRUE              TRUE       TRUE
#> 7  010101-0011   FALSE       FALSE            TRUE             FALSE       TRUE
#>   valid.day valid.month valid.year valid.length valid.century
#> 1      TRUE        TRUE       TRUE         TRUE          TRUE
#> 2      TRUE        TRUE       TRUE         TRUE          TRUE
#> 3      TRUE        TRUE       TRUE         TRUE         FALSE
#> 4     FALSE        TRUE       TRUE         TRUE          TRUE
#> 5      TRUE       FALSE       TRUE         TRUE          TRUE
#> 6      TRUE        TRUE       TRUE        FALSE          TRUE
#> 7      TRUE        TRUE       TRUE         TRUE          TRUE