Finnish personal identification number extraction

Extract information from Finnish personal identification numbers (hetu).

hetu(pin, extract = NULL)

Arguments

pin
Finnish personal identification number as a character vector, or vector of identification numbers as a character vectors

extract
Extract only selected part of the information. Valid values are "hetu", "gender", "personal.number", "checksum", "date", "day", "month", "year", "century.char". If NULL (default), returns all information.

Value

Finnish personal identification number data.frame, or if extract parameter is set, the requested part of the information as a vector. Returns NA if the given character vector is not a valid Finnish personal identification number.

hetu

Finnish personal identification number as a character vector.

gender

Gender of the person as a character vector ("Male" or "Female").

personal.number

Personal number part of the identification number.

checksum

Checksum for the personal identification number.

date

Birthdate.

day

Day of the birthdate.

month

Month of the birthdate.

year

Year of the birthdate.

century.char

Century character of the birthdate: + (1800), - (1900) or A (2000).

See also

Examples

hetu("111111-111C")

#>          hetu gender personal.number checksum       date day month year
#> 1 111111-111C   Male             111        C 1911-11-11  11    11 1911
#>   century.char
#> 1            -

hetu("111111-111C")$date

#> [1] "1911-11-11"

hetu("111111-111C")$gender

#> [1] "Male"

# Same as previous, but using extract argument
hetu("111111-111C", extract="gender")

#> [1] "Male"

# Process a vector of hetu's
hetu(c("010101-0101", "111111-111C"))

#>          hetu gender personal.number checksum       date day month year
#> 1 010101-0101 Female              10        1 1901-01-01   1     1 1901
#> 2 111111-111C   Male             111        C 1911-11-11  11    11 1911
#>   century.char
#> 1            -
#> 2            -

# Process a vector of hetu's and extract gender information from each
hetu(c("010101-0101", "111111-111C"), extract="gender")

#> [1] "Female" "Male"