Finnish personal identification number extraction

Extract information from Finnish personal identification numbers (hetu).

hetu(pin, extract = NULL, allow.temp = FALSE, diagnostic = FALSE)

Arguments

pin
Finnish personal identification number as a character vector, or vector of identification numbers as a character vectors

extract
Extract only selected part of the information. Valid values are "hetu", "sex", "personal.number", "checksum", "date", "day", "month", "year", "century", "is.temp". If NULL (default), returns all information.

allow.temp
Allow artificial or temporary PINs (personal numbers 900-999). If FALSE (default), only PINs intended for official use (personal numbers 002-899) are allowed.

diagnostic
Print additional information about possible problems in PINs. The checks are "invalid.personal.number", "invalid.checksum", "incorrect.checksum" "invalid.date", "invalid.day", "invalid.month, "invalid.length, "invalid.century". Default is FALSE which returns no diagnostic information.

Value

Finnish personal identification number data.frame, or if extract parameter is set, the requested part of the information as a vector. Returns an error or NA if the given character vector is not a valid Finnish personal identification number.

hetu

Finnish personal identification number as a character vector. A correct pin should be in the form DDMMYYCZZZQ, where DDMMYY stands for date, C for century sign, ZZZ for personal number and Q for checksum character.

sex

sex of the person as a character vector ("Male" or "Female").

personal.number

Personal number part of the identification number.

checksum

Checksum for the personal identification number.

date

Birthdate.

day

Day of the birthdate.

month

Month of the birthdate.

year

Year of the birthdate.

century

Century character of the birthdate: + (1800), - (1900) or A (2000).

is.temp

Is the personal identification number an artificial number intended for temporary use: (TRUE or FALSE)

See also

Examples

hetu("111111-111C")

#>          hetu  sex personal.number checksum       date day month year century
#> 1 111111-111C Male             111        C 1911-11-11  11    11 1911       -
#>   is.temp
#> 1   FALSE

hetu("111111-111C")$date

#> [1] "1911-11-11"

hetu("111111-111C")$sex

#> [1] "Male"

# Same as previous, but using extract argument
hetu("111111-111C", extract="sex")

#> [1] "Male"

# Process a vector of hetu's
hetu(c("010101-0101", "111111-111C"))

#>          hetu    sex personal.number checksum       date day month year century
#> 1 010101-0101 Female             010        1 1901-01-01   1     1 1901       -
#> 2 111111-111C   Male             111        C 1911-11-11  11    11 1911       -
#>   is.temp
#> 1   FALSE
#> 2   FALSE

# Process a vector of hetu's and extract sex information from each
hetu(c("010101-0101", "111111-111C"), extract="sex")

#> [1] "Female" "Male"